Add resource economy: barracks upkeep, archer unit cost, player-built barracks
Three changes that turned out to be one connected piece of work once
I started on them:

1. Archer barracks now continuously drains electricity - 0.5/sec idle,
   2/sec while its queue has anything in it. Unconditional, doesn't
   care whether the player can afford the unit being built (that's a
   separate check) - clamps at 0 rather than going negative.

2. The archer unit itself costs 300 iron + 300 electricity. Spread
   evenly across UNIT_BUILD_TIME rather than charged as a lump sum, so
   it reads the same "everything in units/sec" way the barracks upkeep
   does: cost * deltaTime / UNIT_BUILD_TIME per tick. If a tick can't
   afford its share of either resource, production progress just stalls
   (same pattern already used for the MAX_TOTAL_UNITS cap) rather than
   losing progress or draining into the negative. Warrior stays free
   (0/0) - home doesn't check or consume anything.

3. Player resources are float now, not int - 0.5/sec has nowhere to
   accumulate in a whole-number field otherwise. Nice side effect:
   ResourceExtractionSystem and ResourceExtractorComponent both got
   simpler, since the fractional-progress accumulator they used to need
   (to eventually credit a whole unit) isn't necessary anymore - they
   just add rate * deltaTime directly now. Client still displays the
   resource panel as a truncated integer; only the internal accounting
   needed the precision.

Power plant output raised 1 -> 150/sec, since 300 electricity at the
old rate would've taken five minutes to save up for one archer. Iron
mine's rate is untouched - only power plant was asked for - which
leaves it as a likely bottleneck at the same 300-iron cost (worth
revisiting, flagged in the README rather than silently changed).

The archer barracks itself stopped being server-spawned - now it's a
third option in the build menu alongside the mine and the plant, same
free-placement rules as the power plant (BuildingPlacement.canPlaceBuilding,
generalized from the power-plant-only canPlacePowerPlant it used to be).
Gave it buildTime=10 to match the other two player-built buildings,
since instant-appear stopped making sense once a player places it
themselves - not something explicitly requested, calling it out as my
own inference. PlacePowerPlantRequest is gone, replaced by one general
PlaceBuildingRequest{buildingType, x, y} that both the barracks and the
plant send. GameScreen's placingBuildingType switched from ResourceType
to BuildingType, since the barracks isn't a resource building - the old
type didn't have room for it.

Home is now the only building the server spawns automatically on join.
Pathfinding's static obstacle grid dropped the barracks accordingly -
same reasoning already applied to the mine and the plant: it's placed
at runtime by the player now, not at a fixed position known at class
load.

+++ shared/src/main/java/ru/socol/supreme/shared/network/messages/PlayerResources.java
@@ -8,15 +8,19 @@ package ru.socol.supreme.shared.network.messages;
  * ProductionComponent, который тоже служит и серверным состоянием, и
  * тем, что видит клиент.
  *
- * Пока ничего не добывает и не тратит эти числа — зданий добычи ещё нет
- * (см. ResourceType) — оба поля всегда 0. Само хранение и рассылка уже
- * готовы, чтобы будущим зданиям добычи было куда писать.
+ * float, а не int: потребление и добыча считаются непрерывно, в
+ * единицах/сек (ResourceExtractionSystem, ProductionSystem — здание
+ * добычи или казарма стрелков могут тратить/добывать и дробную величину
+ * за тик, например 0.5 электричества/сек в простое у казармы). Клиент
+ * показывает игроку округлённое до целого значение (GameScreen
+ * .drawResourcePanel) — дробная точность нужна только для внутреннего
+ * счёта, не для интерфейса.
  */
 public class PlayerResources {
 
     public int playerId;
-    public int iron;
-    public int electricity;
+    public float iron;
+    public float electricity;
 
     public PlayerResources() {
     }
